Move last parts of Initialized and Initializing into wallet
